Keeping two versions of the docs, latest and helm.
We want to keep a version of the older Helm documentation around for
users still running Helm. We also want all the old links to work so the
SEO doesn't break. So this effectively mirrors the old docs and
redirects until such time as we have content for these sections.